requirements.txt
# ===========================================
# Interview Helper - Core Dependencies
# ===========================================

# Web Framework
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.6

# AI and ML
openai==1.97.0
torch==2.7.0+cu118
transformers==4.51.3
sentence-transformers==4.1.0

# Data Processing
numpy==2.2.5
pandas==2.2.3
scikit-learn==1.7.0

# Database
sqlalchemy==2.0.41
pydantic==2.11.7

# Environment
python-dotenv==1.0.0

# File Handling
aiofiles==23.2.1
PyPDF2==3.0.1

# HTTP and WebSocket
requests==2.32.4
websockets==15.0.1